requirements.txt
django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf_yasg==1.21.7
pillow==10.2.0
python-dotenv==1.0.0
requests==2.28.1